When fetching feeds, set the feed string's encoding to ISO-8859-1 if necessary.
This happens when fetching a feed that does not specify an encoding. In this case the HTTP spec says that the string should be assumed to be encoded in ISO-8859-1, but RestClient returns a string that reports its encoding as UTF-8, even if it's actually not valid UTF-8 (it's ISO-8859-1). The fix in this case is to tell the string to report its encoding as ISO-8859-1 before operating on it.