Added a "time" request parameter to all AJAX GET requests, with the current time in milliseconds since Epoch, to bust XHR caching.
Before this change, the URL used in AJAX GET request to retrieve e.g. the list of feeds was always the same. This meant that some browsers cached the response and the first such AJAX request was actually served from the cache instead of being sent to the server. The result was that when the user opened the browser after e.g. several hours closed, at first the feeds and unread counts being shown actually were those from last time, which the browser had cached; only after one minute an actual HTTP request happened and the actual feeds and counts were displayed. This was confusing for users.

The workaround is to include a parameter in each AJAX GET which changes every time. The current time in ms after Epoch (1/1/1970) has been chosen for simplicity and because it is more or less guaranteed to change on every request. This means that the URL being used every time is different, and the browser no longer caches responses. This "time" request parameter is ignored by the server.