Updated angularJS from 1.0.8 to 1.2.6
From version 1.2 onward, ng-bind-html-unsafe (which feedbunch used to bind feed titles, entry summaries and the like into the view) has disappeared. The ng-bind-html directive is to be used instead. This requires marking the strings we want to bind as html-safe, which is burdensome code, or making the angular-sanitize module available to our angularjs module (if none of those things is done, errors appear in the js console and the "unsafe" and "unsanitized" strings are rejected, which means lots of strings missing from hte view). Including angular-sanitize in the app is the easiest migration path.

For more details about this change see angularjs release notes for version 1.2.0:

  https://github.com/angular/angular.js/blob/master/CHANGELOG.md

Sanitizing these strings is a bit redundant because they are alredy sanitized by the rails models before saving them in the db, but for now the performance hit seems to be imperceptible.