The start page now shows the number of subscribed feeds, regardless of whether they are currently displayed or not (feeds may not be displayed if they currently have no unread entries).
Also the number of subscribed feeds displayed is updated when subscribing and unsubscribing from feeds.

Added acceptance tests to test this behavior, and also to ensure that the count of unread entries in the start page is properly displayed when entries change state (this was already working but it was not being tested).