When loading entries, reset the feeds update timer only if the load method is actually going to do some work.
Before this commit, the timer reset happened at the start of the load_entries function, before the validations that have to pass for the function to actually do some work. This meant that even if the load_entries function was invoked but was not going to actually do anything, e.g. if the next page of entries for the currently selected feed is still loading, the timer was reset anyway. This behavior was not correct.