When loading entries from the server via AJAX, do not run the error callback if the AJAX call is prematurely cancelled (e.g. because the user clicks on another feed while the first one is still loading).
Before this commit the error callback was also executed when the AJAX call was prematurely cancelled, which had the side effect of making the client believe that no AJAX error was currently running. This could cause several AJAX calls for entries to be running in parallel, which gave weird results.