When loading the first page of entries from a feed or folder, a second page of entries is immediately loaded. Additional pages are loaded when the user scrolls down.
This is necessary because ng-infinite-scroll loading of entries is only triggered when the user scrolls down. In the case in wich the viewport is so tall that there's enough height for a page of entries (25 entries) with blank space below, there's no scrollbar and therefore the user has no way to trigger the loading of additional entry pages.

This problem is alleviated by always loading two pages when clicking on a feed or folder. This should be enough for the entries list to be taller than the viewport in most cases, which ensures the presence of a scrollbar and makes the infinite scroll work as intended. In some corner case of freakishly tall viewports I guess the problem could reappear; we'll cross that bridge when we get to it.