Adjusted the point to which the viewport scrolls automatically when opening an entry, so that (if possible) just one entry above the opened one is visible.
With recent changes (entry-list links have less height now) the viewport was scrolling to a point in which the current entry, the one above, and part of the one yet above were visible under the navbar. With this commit the entry two positions above the currently opened one is no longer visible, which is a better use of vertical space.