CSS files are now pure SCSS files. This enables sharing of variables defined in one file across all SCSS files.
Previously these files had the .css.scss extension, which meant that they were converted from SASS to CSS individually before requiring them from the application.css.scss manifest. This meant that each file that used variables defined in _responsive.css.scss had to @import the file in which variables were defined.

Besides being very un-DRY, this also meant that a copy of the _responsive.css.scss file was inserted at the top of EVERY file that was including it, adding to the total CSS size unnecessarily (at least this guy says this is what was happening: http://pivotallabs.com/structure-your-sass-files-with-import/).

Now all stylesheets have only the .scss extension, except the application.css.scss manifest; and stylesheet files are included from the manifest with SASS @include directives, instead of the asset pipeline's require instruction. This means that SASS works as intended, sharing variables and mixins across all files; a small change is that even in development stylesheets are served as a single file, instead of individually (the SASS preprocessor creates a big file with all the @include's, instead of serving individual files in deployment like the asset pipeline does). Still I don't think that will affect me in the least.