If an entry is found with a feed_id which does not correspond to any feed in the scope, immediately trigger a refresh of feeds from the server.
This can happen when retrieving entries from a folder or for the "all subscriptions" link. In this case some entries may be returned for a feed which is not in the scope, because in the last feeds refresh didn't have unread entries (but in the meantime has been refreshed and unread entries have been fetched). In this case an immediate refresh of feeds should retrieve (among others) this new feed. The feed title for the entry will be blank for the user for a few moments until the feeds are loaded from the server, and then the feed title will appear.