When retrieving entries, they are ordered by published date first, creation date in the database second, and id third.
The change in this commit is adding back the id, to break ties for entries which have the same published date and creation date.

This is intended for the case in which two entries are created in such a short interval that the timestamps are the same. In this case we hope the id will be a sequentially increasing number which will break the tie.