Added "last_fetched" and "fetch_interval_secs" attributes to feed model and database table.
The "last_fetched" attribute will store the last time the feed was fetched. It can be nil if the feed has never been fetched (it's just been created).

The "fetch_interval_seconds" will store the current interval between fetches, in seconds. When a feed is created it defaults to 3600 seconds (1 hour). In the future this interval will change in an adaptative way so that feeds that update more often have shorter intervals between fetches, and vice versa.