When updating a feed, if an error is raised, increase the interval between updates (the same as if the update had not failed but had returned no new entries).
Now error raised while fetching the feed do not cause the job to fail. Besides incrementing the update interval, this change ensures that the next update is correctly scheduled. The error is logged.

This can happen if the server is not listening (causing a timeout), if the server responds with a 404, etc. All these are error conditions outside the control of our app; this app should react accordingly, but they should not be treated as errors within the app itself.