Most (hopefully, all) errors that can be raised while updating a feed which are caused by circumstances outside our control (feed server is down, is no longer available, feed URL has changed...) are rescued and logged instead of marking the job as failed.
When one such error is raised, the fetch interval for the feed is increased accordingly (this was already done with errors raised by RestClient, now this behavior has been extended to other errors that can also be raised without actually indicating an error in Feedbunch).

In the future, when a feed has been failing continously for a very long time, it will be marked as permanently bad and no more updates will be attempted.