Added "available" attribute to Feed model and database table.
This attribute is required and defaults to true.

Not yet implemented: only feeds with available=true will have scheduled updates. When a feed's scheduled updates have been failing for a certain period of time (configured in application.rb), the feed will be marked as unavailable (feed.available = false) and no more updates will be attempted.

This is intended for the case in which a feed is no longer available at its fetch_url (e.g. the server disappears, the feed URL changes, the DNS domain expires...), so that Feedbunch does periodically try to fetch from a non-existing server, forever.