Fixed bug in Feed#available default value: when setting it to false, the model was resetting it to true.
This was because the rails .blank? helper was used to detect if the attribute had no value. The way it works, it detects a false boolean value as blank, which triggered the code that sets the default true value.

The fix is to use the .nil? helper instead.

Also the FactoryGirl factory for the Feed model gives an available value of true.