WIP: display useful information in ActiveAdmin views.
Now the user show view displays, below the user attributes, two tables:

- Folders owned by the user
- Feeds subscribed by the user

Also useful information which is not a database column, like number of feeds in a folder, number of unread entries in a feed for the selected user etc are displayed. Each table row has a "View" link which goes to the individual feed or folder show view.

To enable this change some ActiveAdmin views are rendered with partials (files with .builder extension in app/views/admin subfolders) instead of configuring the view in the ActiveAdmin resource definition (files under /app/admin).

The change to partials has been inspired by this StackOverflow post:

http://stackoverflow.com/questions/13422751/rails-activeadmin-showing-table-of-a-related-resource-in-the-same-view