Added ID to all resource views in ActiveAdmin.
It's a good idea to have easy access to the id of model instances. E.g. if a Resque job is failing, the only clue we have of which feed is failing is the ID passed as argument to the job; with it, now we can search in ActiveAdmin the specific feed and see all its details.