WIP: lazy loading images.
It's starting to look like the best option is to implement this in an angular service, instead of using js libraries. This would allow other functionalities to be added (adding a class with margins and display-block only if the image is big enough, try to load the image as a relative URL in the feed server if the first attempt to load it fails).

Not yet implemented:

- remove all traces of jquery-unveil library (no longer used)
- move lazy-loading code to its own angularjs service
- implement the above additional features
- implement a lazy-loading solution when user has requested all entries to be open by default