Remove margins from "mark as read" and "mark as unread" buttons in the entry toolbar.
Bootstrap by default gives a 5px margin-left to the second and successive buttons in a toolbar. We are showing only one button at once, which meant that the button that appeared first in the markup was displayed with no margin-left, and the second one was displayed with a small margin-left. The visual effect was that the button "jumped" a bit to the left or right when clicking.

The fix is to remove margins for those buttons. This will probably require some refining when more buttons are added (e.g. social sharing).