Tooltips are not enabled in smartphone and tablet-sized screens.
Tooltips do not make much sense in touchscreens, in which there is no hover. The media query used matches the $screen-md-min cuttoff point for Bootstrap; this should make the app behave coherently at different widths. There is the possibility that a user with a mouse, who would be interested in tooltips, could have a viewport under this width, but it's something of a fringe case which cannot be easily addressed.

To enable media queries in Javascript the Enquire.js library is used:

	http://wicky.nillia.ms/enquire.js/