The animated GIF displayed while entry images are lazily loading is in the public folder, outside the asset pipeline.
This is necessary because we need a static path for this particular image. The path to the image is inserted in the markup of entries when they are saved in the database; this means that we cannot use the asset pipeline, because the image path could change when deploying (e.g. if we change the image with a different one); this is generally desirable, because views are generated on the fly to point to the correct asset path (modified by the asset pipeline). But in the case of entries, they are not rails views but rather markup fragments saved in the database which once saved are never changed, and we need that images referenced there (e.g. the animated gif with which we replace images for lazy-loading) can be downloaded at the same path that when the entry was originally saved.