Do not hide images while they are lazy-loading.
It is a good idea to give the user a visual indication that something is loading, so that he's not surprised when an image appears in the middle of the text he was reading. This is specially true with small network speeds. Showing an animated spinner until the image is fully loaded is better than showing nothing.