Fixed errors caused by recent updates to Feedjira and Rubyzip gems.
Specifically,

- Zip::ZipError is now called Zip::Error
- Feedzirra::Feed is now called Feedjira::Feed

Also added a short sleep to a feature test to get rid of occasional failures (sometimes not all data had been properly saved in the db before the test tried to mark all entries as read). I don't like it and I'll probably try to get rid of the sleep in the future, but for now I've found no other reliable way for the test to consistently pass.