In smartphone-sized screens, when the main menu is automatically collapsed, give focus to the main "Feedbunch" brand link in the navbar.
This is intended to complement commit 34237740db7f26f25fc3b8f0f34fdf211484b4ea, so that the last selected menu entry does not retain focus (and therefore a dark background which works as a highlight) after doing its work. It was surprising for the user to open the menu again and see highlighted the last entry menu he clicked on.