In smartphone-sized screens, when collapsing the menu, give focus to the toggle-menu-collapse button in the navbar.
Related to 1d5e4d0910442c69bdad11fd73b2b27852f8e8e8. The only change is which element is focused; the reason is that Android browsers give focused elements a weird orange highlight which looks weird around the brand link in the navbar. This hightlight is not straightforward to remove, and it's simpler to give focus to a related button which does not look so bad.