In smartphone-sized screens, when collapsing the main menu, remove focus from all links in the navbar (and menus under the navbar).
This is related with commit b89ca966b257d431db503a8d249ba6aa21cace85. It happens that giving focus to the menu-toggle button when collapsing the menu is not good enough, because certain user actions can trigger this and give focus to the button unexpectedly (e.g. simply opening a folder). It's better to remove focus in this case.