DataImport model now accepts "NONE" as a valid state. This is now the default state when creating a new DataImport instance.
Until now, when a user has never ran an OPML import there is no DataImport instance in the DB for that user. With the changes being implemented in this and the next few commits, all users will have an associated DataImport record, which will have state "NONE" for those users who have not ran an OPML import.