All users now have an associated DataImport instance as soon as they are created, with status "NONE".
The view that generates JSON for the DataImports no longer has to make a special case for when the user has no DataImport, this should never happen.

Added db migration to create missing DataImports for those users that have not yet ran an OPML import.

WIP: some tests currently failing.