Large images in entries have a 1em top and bottom margin. Small images in entries have no top or bottom margins, but small 3px right and left margins instead.
The reason is that small images are sometimes in the middle of a line of text, or they have their text description right after the image. In these case it looks bad if those small images have a bottom margin, because the image is one line above the surrounding text which looks weird. It's generally better that small images have no bottom/top margins. A small right/left margin is desirable in case the image itself has no white around it, to separate it from the surrounding text.