Changed all routes which accept or return JSON to a named "api" subpath.
This means instead of /feeds.json, now /api/feeds.json will be invoked.

This means a change in controllers, views, routes and most javascript that uses $http to send and receive JSON via AJAX.

This way project has better maintainability. E.g. all views that generate JSON and are meant to be invoked via AJAX are under a /app/views/api folder. Also this would enable us in the future to support a new version of the API e.g. in the "apiv2" subpath without removing the current API.