Added explicit return statement to most AngularJS controller methods (except those that already did return a value).
This is necessary so that Coffeescript does not return the value of the last method line as the method's return value; in some cases this could be a jQuery node which would then be inserted in the angularJS view expression that invoked the method. This (directly using DOM nodes in AngularJS expressions) is explicitly forbidden since angularJS 1.2, and raises a "[$parse:isecdom]" error.

For more about this odd error see:

https://groups.google.com/forum/#!msg/angular/bsTbZ86WAY4/hh_319IrmFoJ