Explicitly configure UTC timezone for the app. Replaced DateTime.now with Time.zone.now and DateTime.new with Time.zone.parse.
DateTime.now and DateTime.new produce DateTime instances which have no timezone information; more specifically, they are in the timezone currently configured for the server. This dependence on the server configuration is generally undesirable.

It's better to explicitly configure a timezone and use Time.zone.now and Time.zone.new, which are replacements that generate times (and dates) with timezone information (specifically the timezone configured in application.rb). The app has timezone information saved for each user, and converts times (entries publish dates, mostly) to the user timezone before displaying them in the browser.

Actually configuring the timezone in application.rb is not mandatory, because Rails defaults to UTC timezone, but I think it's clearer to specify it in the config.