Moved the code that updates a feed to a new class lib/feed_updater. UpdateFeedJob just invokes this class.
This enables us to create a different job for when the update is manually requested by a user. This job will receive as an additional argument the id of the user, will also invoke lib/feed_updater and additionally will update the status of the job in the database, so that the user can see this status in the view.