WIP: display status of refresh jobs in the start page, and display an alert when such a job finishes.
I've decided not to return full data for the related feeds when returning a JSON document with the refresh_feed_job_statuses. It's too verbose and in many cases redundant; it will be the responsibility of the client to request the information of any feed which is not already in the scope.