When setting the current feed or folder in the angularjs app, hide the sidebar instead of toggling its current visibility state.
When the only way to set the feed/folder was clicking on the sidebar links, toggling worked fine because for this to happen the sidebar had to be visible, and toggling made it hidden. However now the current feed can also be set by clicking on the feed link in a job status alert; when this happens the sidebar is hidden, and toggling it would make it shown, which is the opposite of the desired behavior. The fix is to hide it unconditionally.

The fix is only really necessary when setting the current feed, but it's also applied when setting the current folder for coherence in the code.