When setting the current feed or folder in the angularjs app, open the corresponding folder if it isn't already open.
This is necessary when the user clicks on the feed link in a job status alert. This sets the current feed to the refreshed feed, but if the feed's folder is not open in the sidebar, it is not immediately obvious that that feed has been selected.

The fix is to open automatically the folder (and close any other open folder), so that when the user clicks on the link the feed becomes automatically visible in the sidebar and it's obvious it's been selected.