The FeedsController#create action does not immediately subscribe a user to a feed anymore.
Instead, it invokes a new User instance method which enqueues a job that will perform the actions necessary to subscribe the user. Also a SubscribeJobState instance is created, so that the user can be informed of the state of his subscription request.

If the user is already subscribed to a feed with the same fetch_url, no job is enqueued.