Complement the changes in 47b69ebdcf6e48abc02a69e9ed447e8c3d5e6703 making the corresponding changes in the angularjs client.
Obtaining user config and obtaining other user data has been separated into two different angularjs services. Both are invoked when initializing the app, but after successfully subscribing to a feed only user data is loaded again, to get the new count of subscribed feeds (which is displayed in the start page).

This fixes the circular dependency error, hopefully for good this time.