Added one-to-many association between feeds and subscribe_job_states.
If the subscribe_job_state has state "SUCCESS", its feed_id foreign key must be different from null. If the job has any other state, the feed_id foreign key must be null.

This makes sense because until a subscribe job is finished successfully, we cannot be sure that the associated feed exists in the database.