WIP: resourceful route to get a list of feeds in a folder.
The angularjs client will use it to check if a just unsubscribed feed was the only feed in its folder; in this case, the folder will be removed from the dropdown menu.

This is necessary because the actual unsubscribing no longer happens immediately but is rather done in a background job; therefore the folder will not be destroyed until the job is processed, which may take some time. Despite this we want to give the user the impression that the unsubscribing and the folder being destroyed happen immediately: fake it till you make it!