When unsubscribing, if there were no more feeds in the folder besides the unsubscribed feed, the folder is immediately removed from the scope.
This affects both folders in the sidebar and the folders dropdown.

Since unsubscribing is not immediate but rather happens after a certain delay in a background job, the angularjs client requests from the server a list of feeds in the folder, and if the only feed returned is the unsubscribed one, the folder is removed from the client (the actual removal from the db will happen when the job is performed). This is a case of "fake it till you make it" which should feel responsive to users in most cases.