When unsubscribing, request all feeds in the folder when trying to determine if the folder should be removed, instead of requesting only feeds with unread entries in the folder.
If there are other feeds in the folder besides the just unsubscribed one, the folder should be present in the scope regardless of whether those feeds have unread entries or not.

This is because all folders should appear in the dropdown, so that the user has the option of moving any feed into any of his existing folders. The sidebar is another story: only folders and feeds with unread entries should appear there (unless the "show read" button has been clicked), but the view takes care of filtering and hiding folders in the sidebar, instead of removing them from the scope.