Solve timing problem in acceptance test.
Instead of just reloading the page, which can be so quick that the subscribe_job_state has not yet been deleted from the db, the user is logged out and logged in again. This should give enough time for the subscribe_job_state to be deleted from the db, which should make the alert disappear from the start page.