User#folder_feeds now accepts as argument Folder::NO_FOLDER as well as a Folder instance.
If this value is passed to the "folder" argument, the method returns the list of subscribed feeds which are not in any folder. The optional include_read argument is supported as before (it controls whether all feeds are returned, or only feeds with unread entries).

The old behavior (passing a Folder instance makes the method return the list of feeds in the folder) works as before.