Moved some of the markup for the opml_import_job_state alerts from the start partial to the opml_import partial view.
The ng-show attribute that controls if the alert is visible or not was in the start view, which didn't make much sense (it is functionally related to the opml import alerts, and one would expect to find it in that partial). Not that ng-show is in the opml_import partial.