Fixed code that detects if an OPML import has finished, based on the presence of SubscribeUserJob instances still enqueued for the user.
SubscribeUserJob checks if an OPML import has finished by, among other things, looking at the jobs queue checking if there are other SubscribeUserJob instances enqueued for the same OPML import.

This code had stopped working because it was looking in the "update_feeds" queue, but recently SubscribeUserJob has been changed to run in the "subscriptions" queue.

The fix is to look for enqueued instances of SubscribeUserJob in the "subscribe" queue, instead of the "update_feeds" queue.