Added callback that delestes any associated OPML file before destroying an OpmlExportJobState.
This means that OPMLExporter.export no longer needs to explicitly delete old OPML files when doing a new OPML export. This is automatically taken care of by User#export_subscriptions, when it creates a new opml_export_job_state and deletes the old one (which triggers the deletion of the old OPML file, if any).