The name of the saved OPML export files is always "feedbunch_export.opml". Those files are organized in subfolders according to the user ID, so that there is no possible confusion as to whom belongs a given export file.
Also now that imports files are also organized in subfolders by user ID, the chance of collision is negligible. The naming method uses the current time in seconds; until recently it was possible two users uploaded a file at the same moment, causing a funny situation in which the app behaved incorrectly. Now the same situation could happen but a user would have to upload two OPML files within a 1-second interval. This is not really gonna happen during normal use.