When the OPML export is available for download, show the date in which it was generated. Also show the download link permanently in the edit registration view.
Showing the date will help the user decide if the export is too old and he has to generate a new one.

Showing the link in the edit registration view means that the user, once he has finished an export, will always have the download link available (the link in the start page alert may disappear if the user clicks on the "dismiss" button).