Refined the angularjs condition that makes the opml export alert appear in the start page.
Before this when the export had never ran (the OpmlExportJobState had state "NONE"), an empty div appeared in the page which made a certain acceptance test often fail.

Now for the div that contains the alert to appear, both the state has to be different to "NONE" and the boolean show_alert has to be true (this is, the user must never have manually closed the alert).