WIP: user must enter his password to cancel his account.
Implemented: when clicking on "delete account" button in the edit registration view, instead of a simple "accept/cancel" confirm dialog a bootstrap popup is displayed, including a form with a field for the password.

Not yet implemented: validate that the password entered is correct. Change acceptance tests in sync with these changes.