Added the correct type attribute to buttons in the form-footer of the delete profile form.
Specially important is that the close button has type="button". Otherwise some browsers (at least firefox, haven't tested others) gets confused and think that the close button is actually the submit button. This results in the following bug: when the user writes the password and hits Enter key, the browser activates the Close button (thinking it's the submit button for the form), which actually dismisses the popup without submitting the form.

By adding the correct type attributes, the form is submitted when hitting the enter key with the password field focused, as expected.