Account deletion now happens in a background job.
When a user requests his account to be deleted, a job is enqueued to perform the deletion. Also the user account is immediately locked so that he cannot log in before the job is performed and the account actually deleted.

This should improve user experience when a user with many subscriptions deletes his accounts. Actually deleting the account takes some time, but by delegating it to a background job the user gets an immediate response.