When a feed has a subscribe alert visible in the start page, it is never removed from the sidebar even if it has no unread entries.
This was already implemented for the case in which the feed has a refresh alert in the start page, now it has been extended for the subscribe alert. There are currently no other alerts that can appear in the start page which have a link to a feed.

This is intended so that the feed is in the scope, even if it has no unread entries, and the link displayed in the subscribe or refresh alert is functional (when it is clicked, the user is shown the entries from that feed).