When retrieving paginated entries, stop asking for more pages when a page with less than the maximum number of entries is received.
Previously, the next page was always requested from the server, until a 404 was received. This may still happen when the number of entries is an integer multiple of the page size, but in most cases the client will stop asking for more pages when the last page is retrieved; the client can assume that if less than the maximum amount of entries are received in a page, this is the last page and it is not necessary to ask for the next page of entries.