When retrieving feeds in a paginated fashion, if a page with less than the maximum number of feeds per page is retrieved, stop paginating.
Previously the next page of feeds was requested, until the server returned a 404. Now if the client receives less than the maximum number of feeds in a page it can assume that it is the last page, and the last feeds page request (which most of the time resulted in a 404) is avoided most of the time. It can still happen if the number of feeds is an integer multiple of the maximum number of feeds per page.