WIP: automatically delete old entries, so that database size does not grow indefinitely.
Currently implemented: created DeletedEntry model; the idea is that when entries are deleted in an automated cleanup, a DeletedEntry instance is saved in the db. This will be used in the future to ensure that an already deleted entry is not saved again in the database.

This means that deleted entries will occupy some space in the database, and database size will grow over time. However each deleted entry will need very little database space (specially because the entry content, which is by far the biggest field, won't be kept) and the growth should be manageable. Also, the read/unread states for deleted entries won't be kept, which also means some space saving.