Automatically delete old entries.
Every time a scheduled feed update is run, after inserting any new entries, the number of entries in the feed is counted to check if it is above the limit configured in application.rb. If it is, older entries are deleted until the number of entries in the feed is the maximum allowed.

By "delete entries" I mean that the Entry record is deleted in the database, and a corresponding DeletedEntry record is inserted in the db so that the entry will not be saved again in the database, even if the feed server returns it again in the future. Destroying Entry instances also triggers the destruction of associated EntryStates.

This should keep the database from growing too fast, keeping old entries indefinitely does not add much value.