Defensive programming: when the OPML export job has state "SUCCESS", guard against the case in which export_date attribute is set to nil.
This should be very rare, because validations should ensure that export_date has a valid value when the state of the OPML export job is set to SUCCESS. However, in the odd case that export_date is nil (e.g. old db records from before export_date was added to the table), updated_at is used instead. I trust ActiveRecord will take care of giving value to that field in all cases.